refactor: follow mvc pattern
