+++ src/routes/studyplans.route.ts
@@ -1,0 +1,8 @@
+import express from "express";
+import { getAllStudyPlans } from "../controllers/studyplans.controller";
+
+const studyplanRouter = express.Router();
+
+studyplanRouter.get("/", getAllStudyPlans);
+
+export default studyplanRouter;
