+++ src/utils/converter.ts
@@ -1,5 +1,5 @@
 import { WithId, Document } from "mongodb";
-import { StudyPlan } from "../model/model";
+import { StudyPlan } from "../model/studyplans.model";
 
 export function convertResultToStudyPlan(result: WithId<Document>): StudyPlan {
     return {
