+++ src/utils/s3connection.ts
@@ -4,6 +4,7 @@ import { S3Client } from "@aws-sdk/client-s3";
 dotenv.config();
 const s3Client = new S3Client({
     endpoint: process.env.S3_ENDPOINT,
+    forcePathStyle: true,
 });
 
 export { s3Client };
